# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MATCHPIT API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get user dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─────────── NOTIFICATIONS ───────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        favoriteSports:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
          nullable: true
        walletBalance:
          type: number
        badgeCount:
          type: integer
        trustScore:
          type: number
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, fullName, email, walletBalance, badgeCount, trustScore, isAdmin, createdAt, favoriteSports]
...
    Venue:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        address:
          type: string
        sports:
          type: array
          items:
            type: string
        pricePerHour:
          type: number
        coverImage:
          type: string
          nullable: true
        rating:
          type: number
        totalReviews:
          type: integer
        isApproved:
          type: boolean
        amenities:
          type: array
          items:
            type: string
      required: [id, name, city, address, sports, pricePerHour, rating, totalReviews, isApproved, amenities]
...
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        venueId:
          type: string
        slotId:
          type: string
        sport:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending_payment, confirmed, cancelled, completed]
        paymentId:
          type: string
          nullable: true
        venue:
          $ref: "#/components/schemas/Venue"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, venueId, slotId, sport, date, startTime, endTime, totalAmount, status, createdAt]
...
    HostedMatch:
      type: object
      properties:
        id:
          type: string
        hostUserId:
          type: string
        venueId:
          type: string
        slotId:
          type: string
        sport:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        totalPlayers:
          type: integer
        minPlayers:
          type: integer
        currentPlayers:
          type: integer
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, any]
        hostFee:
          type: number
        reserveFee:
          type: number
        finalFeePerPlayer:
          type: number
        totalVenueCost:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, confirmed, funded, cancelled, expired]
        financialStatus:
          type: string
          enum: [pending, partially_funded, funded]
        venue:
          $ref: "#/components/schemas/Venue"
          nullable: true
        host:
          $ref: "#/components/schemas/Profile"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, hostUserId, venueId, slotId, sport, date, startTime, endTime, totalPlayers, minPlayers, currentPlayers, skillLevel, hostFee, reserveFee, finalFeePerPlayer, totalVenueCost, status, financialStatus, createdAt]
...
    DashboardSummary:
      type: object
      properties:
        upcomingBookingsCount:
          type: integer
        totalMatchesJoined:
          type: integer
        totalMatchesHosted:
          type: integer
        walletBalance:
          type: number
        pendingFinalPaymentsCount:
          type: integer
        upcomingBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        confirmedMatches:
          type: array
          items:
            $ref: "#/components/schemas/HostedMatch"
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
      required: [upcomingBookingsCount, totalMatchesJoined, totalMatchesHosted, walletBalance, pendingFinalPaymentsCount, upcomingBookings, confirmedMatches, badges]
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [booking_confirmed, match_joined, match_hosted, match_confirmed, payment_success, badge_earned]
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        referenceId:
          type: string
          nullable: true
      required: [id, type, title, description, createdAt]
    Badge:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        label:
          type: string
        description:
          type: string
        icon:
          type: string
        earnedAt:
          type: string
          format: date-time
      required: [id, slug, label, description, icon, earnedAt]

    # ─── NOTIFICATIONS ───
...